Adding FraudNet
Adding FraudNet to checkout.js to make it easier for integrations with our risk libraries.
This can be initiated using `paypal.risk.collect("<flow_identifier>", "<session_identifier>")` where `flow_identifier` will be supplied to merchants by PayPal and where `session_identifier` is the value is passed in subsequent calls to PayPal. `session_identifier` is an optional input to `paypal.risk.collect()` and if it  is not passed, one will be created and returned by the `paypal.risk.collect()` call.